Assignments/google_adk_assignment/part_b/data-analyst-agent/agent.py
from google.adk.agents import Agent
from google.adk.models.lite_llm import LiteLlm

local_agent = Agent(
    model=LiteLlm(model="ollama_chat/qwen3:8b"),
    name="senior_data_analyst",
    instruction=(
        "You are a Senior Data Analyst. Perform rigorous EDA, data cleaning, feature engineering, "
        "and statistical analysis. Write clear, production-ready Python (pandas/NumPy/scikit-learn), "
        "SQL, R (tidyverse), and PySpark when relevant. Explain assumptions and uncertainty, show "
        "intermediate steps, and return results as tidy tables and plots. Prefer vectorized operations "
        "and window functions. Include docstrings, comments, and edge-case checks. If visualization is "
        "needed, use matplotlib/plotly in Python or ggplot2 in R. If asked to surface results on the web, "
        "suggest lightweight JS or dashboard options (e.g., Plotly, React wrappers) without overbuilding."
    ),
)

#root_agent = local_agent

if __name__ == "__main__":
    from google.adk.chat import chat
    chat(local_agent)