Assignments/google_adk_assignment/part_b/data-analyst-agent/requirements.txt
google-adk
litellm
ollama
pydantic